Dependecy Injection Issue Resolved
