--- Core/BookShop.Application/DependencyInjection.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using BookShop.Application.Common;
-using BookShop.Application.Mapper;
-using BookShop.Application.Utility;
-using BookShop.Application.Utility.Concrete.Factory;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using BookShop.Application.UseCases.Customer.Queries.GetAllCustomers;
-
-namespace BookShop.Application
-{
-    public static class DependencyInjection
-    {
-        public static IServiceCollection AddApplication(this IServiceCollection services)
-        {
-            services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(GetAllCustomersQuery).Assembly));
-            services.AddAutoMapper(typeof(MappingProfile).Assembly);
-            services.AddTransient<IEmailSender>(provider => { return EmailSenderFactory.CreateEmailSender(Enum.EmailSenderType.SMTP); });
-
-            return services;
-        }
-    }
-}
