+++ Core/BookShop.Application/UseCases/User/Commands/CreateUser/CreateUserCommand.cs
@@ -18,7 +18,7 @@ namespace BookShop.Application.UseCases.User.Commands.CreateUser
     }
     public class CreateUserCommandHandler : IRequestHandler<CreateUserCommand, bool>
     {
-        public IGenericRepository<BookShop.Domain.Entities.User> _genericRepository;
+        protected readonly IGenericRepository<BookShop.Domain.Entities.User> _genericRepository;
         public IMapper _mapper;
         private IEmailSender _emailSender;
