--- Infrastructure/BookShop.Infrastructure/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace BookShop.Infrastructure
-{
-    public class Class1
-    {
-
-    }
-}
